Adjusted CMake files, buildScripts and tests for building on OS X.  Eliminated compilation of the Navier-Stokes tests until we can substitute an appropriate FFT library (probably FFTPACK).  The build now completes and all tests execute.  5 of 11 tests fail only because the tests do not yet output the expected "Test passed."

+++ src/tests/integration/pde_solvers/coarrayBurgers/global_field.F90
@@ -71,7 +71,7 @@ contains
 
   subroutine set(this,initial_function,num_points)
     class(global_field), intent(inout) :: this
-    integer(int64), intent(in) :: num_points
+    integer, intent(in) :: num_points
     procedure(initial_condition), pointer :: initial_function
     integer(int64) :: num_intervals,i
     real(real64), parameter :: two_pi=2.*3.1415926535897932384626433832795028842_real64
